Add: `acbl_old_cc.sty`
DOES NOT WORK, initial commit.

Adding the old style file.  The intent is to convert it so that
as much as possible of it can be filled out by the new card file.
It won't do everything, it's never intended to - if you play
forcing Major raises or non-Reverse Drury, I don't guarantee
we'll ever be able to check that box.

At first, this will be basic, but I'd like to get it so that
it will "interpret" as much of the new card information and make it
work for the old card.